+++ cmeel/build.py
@@ -1,12 +1,10 @@
-"""Cmeel build.
-
-Functions to generate package archives.
-"""
+"""PEP 517 & 660 entry points."""
 
 import logging
 import os
 
 from .impl import build_impl
+from .sdist import sdist_impl
 
 LOG = logging.getLogger("cmeel")
 
@@ -26,3 +24,8 @@ def build_wheel(wheel_directory, config_settings=None, metadata_directory=None) 
     """Build a binary wheel: main entry point for PEP 517."""
     LOG.info("cmeel build wheel")
     return build_impl(wheel_directory, editable=False)
+
+
+def build_sdist(sdist_directory, config_settings=None) -> str:
+    """Generate a gzipped distribution tarball."""
+    return sdist_impl(sdist_directory)